Modularized all Normalizer into their own shlib.
This is the first step that could lead to Normalizer plugins. It required lots of changes to include directives, but
also it is incompatible with the current state of the logging system. For that reason I gutted the logging for now.
