+++ src/cycle.h
@@ -91,7 +91,6 @@ class Cycle {
           _norm  = true,
           _prob  = true,
           _thread = false;
-     Normalizer::LogLevel _max_log_level = Normalizer::LogLevel::WARN;
      TrainingData* _data;
      Input*  _in;
      Output* _out;
