+++ src/interface/output.h
@@ -31,8 +31,7 @@ class Output : public IOBase {
      virtual ~Output() = default;
      /// put a line on the output device and record it in the history
      virtual void put_line(Normalizer::Result* result,
-                           bool print_prob,
-                           Normalizer::LogLevel max_level);
+                           bool print_prob);
      bool thread_suitable() {
          return true;
      }
@@ -41,8 +40,6 @@ class Output : public IOBase {
      virtual void store_line(const string_impl& line) {
          _training->add_target(line);
      }
-     virtual void log_messages(Normalizer::Result* result,
-                               Normalizer::LogLevel max_level);
      std::ostream *_output;
 };
 
@@ -51,8 +48,7 @@ class Output : public IOBase {
 class InteractiveOutput : public Output {
  public:
      void put_line(Normalizer::Result* result,
-                   bool print_prob,
-                   Normalizer::LogLevel max_level);
+                   bool print_prob);
      std::string validate(const string_impl& line);
      bool thread_suitable() {
          return false;
