+++ src/normalizer/mapper.h
@@ -17,48 +17,5 @@
  */
 #ifndef NORMALIZER_MAPPER_H_
 #define NORMALIZER_MAPPER_H_
-#include<map>
-#include<string>
-#include<fstream>
-#include<tuple>
-#include"string_impl.h"
-#include"base.h"
-#include"result.h"
-
-namespace Norma {
-namespace Normalizer {
-class Mapper : public Base {
- public:
-     void init();
-     using Base::init;
-     void set_from_params(const std::map<std::string, std::string>& params);
-     void clear();
-     Result operator()(const string_impl& word) const;
-     ResultSet operator()(const string_impl& word, unsigned int n) const;
-     bool train(TrainingData* data);
-     void train(const string_impl& word, const string_impl& modern,
-                int count);
-     void save_params();
-     const char* name() const { return "Mapper"; }
-
-     /// Get filename of the current mappings file
-     const std::string& get_mapfile() const { return _mapfile; }
-     /// Set filename for the mappings file
-     Mapper& set_mapfile(const std::string& mapfile) {
-         _mapfile = mapfile;
-         return *this;
-     }
-
- private:
-     ResultSet make_all_results(const string_impl& word) const;
-     bool write_mapfile(const std::string& fname);
-     bool read_mapfile(const std::string& fname);
-
-     std::map<string_impl, std::map<string_impl, int>> _map;
-     std::string _mapfile;
-};
-}  // namespace Normalizer
-}  // namespace Norma
-
+#include"mapper/mapper.h"
 #endif  // NORMALIZER_MAPPER_H_
-
