+++ src/normalizer/mapper/CMakeLists.txt
@@ -1,0 +1,5 @@
+include_directories("${CMAKE_SOURCE_DIR}/src")
+add_library(mapper SHARED mapper.cpp)
+set(NORMALIZER_LIBRARIES ${NORMALIZER_LIBRARIES} mapper PARENT_SCOPE)
+install_headers(mapper.h)
+
