+++ src/normalizer/rulebased/CMakeLists.txt
@@ -1,6 +1,8 @@
-add_sources(rule.cpp rule_learn.cpp
-  rule_collection.cpp candidate_finder.cpp
-  rulebased.cpp)
+include_directories("${CMAKE_SOURCE_DIR}/src")
+add_library(rulebased SHARED
+            rule.cpp rule_learn.cpp rule_collection.cpp
+            candidate_finder.cpp rulebased.cpp)
 install_headers(candidate_finder.h rule.h rule_collection.h rule_learn.h
                 rulebased.h symbols.h)
+set(NORMALIZER_LIBRARIES ${NORMALIZER_LIBRARIES} rulebased PARENT_SCOPE)
